util/requirements.txt
python-dotenv
azure-search-documents==11.6.0b3
azure-identity
openai
azure-cosmos
azure-storage-blob